refactor: move git_cli driven adapter from engine_secret to engine_utilities
