+++ tools/devsecops_engine_tools/engine_sast/engine_code/test/domain/usecases/test_code_scan.py
@@ -6,7 +6,7 @@ from devsecops_engine_tools.engine_sast.engine_code.src.domain.model.gateways.to
 from devsecops_engine_tools.engine_core.src.domain.model.gateway.devops_platform_gateway import (
     DevopsPlatformGateway
 )
-from devsecops_engine_tools.engine_sast.engine_code.src.domain.model.gateways.git_gateway import (
+from devsecops_engine_tools.engine_utilities.git_cli.model.gateway.git_gateway import (
     GitGateway
 )
 from devsecops_engine_tools.engine_sast.engine_code.src.domain.usecases.code_scan import (
